Add tclstub support
To build with tclstub library run.
You have to autogen.sh to create a fresh configure.

NOCONFIGURE=1 ./autogen.sh
./configure --enable-tclstub

And add any other flags you need.

We ship the configure script to help distro's that do not
run a full autogen.sh. But stubs is only available if you
run autogen.sh (or the automake process).