Add user-defined history file name in Setup
The added elseif-branch checks if the environment variable
TCL_HISTORY_FILE is defined and uses it to set historyfile.
The user can define this environment variable by exporting it
in the shell or the shell's rc file, or in .tclshrc with

global env
set env(TCL_HISTORY_FILE_) .app_tclsh_history

This is usefull for applications using an embedded tcl interpreter.
This way, the application-specific extensions, which are only defined
in the application, won't be written to .tclsh-history, which is used
by a normal tclsh session.