Fix pkg load on 8.7
Tcl 8.6 (and 8.5 as far as I can remember) used to infer the name of
this package as Tclreadline.
In Tcl 8.7, the algorithm must have changed (I haven't gone through the
TIPs or the commits). The inferred name is now simply Readline. This
breaks loading, as the init prod is Tclreadline_Init.